Create login-model.js
In this example, we are using the useState hook to keep track of the username and password values as they are entered into the input fields. When the user clicks the "Login" button, the handleSubmit function is called, which can contain your login logic. In this example, we are simply checking if the username and password match hardcoded values, but in a real-world scenario, you would typically make a request to a server to authenticate the user.