Create ajaxpost-model.js
In this example, we are using the fetch API to post the data to a server endpoint. We are also using the useState hook to keep track of the form data as it is being entered. When the form is submitted, the handleFormSubmit function is called, which sends the form data to the server as a JSON payload using the JSON.stringify method. The server response is parsed as JSON using the response.json() method, and the response data is logged to the console. If an error occurs, it is logged to the console using the catch method. Finally, the handleInputChange function is called whenever the user types in an input field, and it updates the formData state object with the new input value.