Create manage-list-model.js
In this example, we are using the useState hook to keep track of the list of items and the new item that the user is typing. We are rendering the list of items as an unordered list using the map method, which iterates over each item and renders a list item element with a "Remove" button. When the "Remove" button is clicked, the handleRemoveItem function is called with the index of the item to remove, which creates a copy of the items array using the spread operator and removes the item at the specified index using the splice method. The setItems function is then called with the updated array to update the state and re-render the list.

To add a new item, we are rendering a form with an input field and a "Add Item" button. When the form is submitted, the handleAddItem function is called, which creates a copy of the items array using the spread operator and adds the new item to the end using the spread operator again. The setItems function is then called with the updated array to update the state and re-render the list. Finally, the newItem state is reset to an empty string to clear the input field.