Create nodejs-db-connect-api.js
Once you have created the API, you can use the fetch method in your ReactJS app to call the API and retrieve the data. Here is an example of how to use fetch in ReactJS
In this example, we are using the useEffect hook to fetch the data from the API when the component mounts. The fetch method sends a GET request to the /data endpoint and retrieves the data as a JSON response. The response data is then stored in the data state using the setData function. Finally, we render the data as a list of items using the map method.